move waitUntilDeploymentReady to DeployWebAppTask
